requirements.txt
#python -m pip install -r requirements.txt

websocket-client
requests
matplotlib
numpy